refactor: nftDownloadHandlers to nft-download.manager
