+++ packages/desktop/electron/lib/managers/index.ts
@@ -2,5 +2,6 @@ export { default as AnalyticsManager } from './analytics.manager'
 export { default as AutoUpdateManager } from './auto-update.manager'
 export { default as DeepLinkManager } from './deep-link.manager'
 export { default as KeychainManager } from './keychain.manager'
+export { default as NftDownloadManager } from './nft-download.manager'
 export { default as NotificationManager } from './notification.manager'
 export { default as PincodeManager } from './pincode.manager'
